pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.6.3",
    "nltk>=3.9.1",
    "numpy==2.1.3",
    "requests>=2.32.5",
    "scikit-learn==1.5.2",
    "scipy==1.14.1",
    "sumy==0.11.0",
]
